+++ src/add-ons/kernel/file_systems/xfs/ShortDirectory.cpp
@@ -86,11 +86,11 @@ ShortDirectory::Lookup(const char* name, size_t length, xfs_ino_t* ino)
 		xfs_ino_t rootIno = fInode->GetVolume()->Root();
 		if (strcmp(name, ".") == 0 || (rootIno == fInode->ID())) {
 			*ino = fInode->ID();
-			TRACE("ShortDirectory:Lookup: name: \".\" ino: (%d)\n", *ino);
+			TRACE("ShortDirectory:Lookup: name: \".\" ino: (%ld)\n", *ino);
 			return B_OK;
 		}
 		*ino = GetIno(&fHeader->parent);
-		TRACE("Parent: (%d)\n", *ino);
+		TRACE("Parent: (%ld)\n", *ino);
 		return B_OK;
 	}
 
@@ -124,7 +124,7 @@ ShortDirectory::GetNext(char* name, size_t* length, xfs_ino_t* ino)
 		strlcpy(name, ".", *length + 1);
 		*ino = fInode->ID();
 		fTrack = 1;
-		TRACE("ShortDirectory:GetNext: name: \".\" ino: (%d)\n", *ino);
+		TRACE("ShortDirectory:GetNext: name: \".\" ino: (%ld)\n", *ino);
 		return B_OK;
 	}
 	if (fTrack == 1) {
@@ -135,7 +135,7 @@ ShortDirectory::GetNext(char* name, size_t* length, xfs_ino_t* ino)
 		strlcpy(name, "..", *length + 1);
 		*ino = GetIno(&fHeader->parent);
 		fTrack = 2;
-		TRACE("ShortDirectory:GetNext: name: \"..\" ino: (%d)\n", *ino);
+		TRACE("ShortDirectory:GetNext: name: \"..\" ino: (%ld)\n", *ino);
 		return B_OK;
 	}
 
@@ -147,7 +147,7 @@ ShortDirectory::GetNext(char* name, size_t* length, xfs_ino_t* ino)
 		uint16 curOffset = B_BENDIAN_TO_HOST_INT16(entry->offset.i);
 		if (curOffset > fLastEntryOffset) {
 
-			if (entry->namelen + 1 > *length)
+			if ((size_t)(entry->namelen + 1) > *length)
 				return B_BUFFER_OVERFLOW;
 
 			fLastEntryOffset = curOffset;
